+++ test/dummy/config/initializers/recording_studio_billing.rb
@@ -2,6 +2,8 @@
 
 require File.expand_path("../../../../app/services/recording_studio_billing/fake_financial_adapter", __dir__)
 require File.expand_path("../../../../app/services/recording_studio_billing/fake_tax_calculator", __dir__)
+require File.expand_path("../../lib/dummy_stripe_checkout_execution", __dir__)
+require File.expand_path("../../lib/dummy_stripe_test_credentials", __dir__)
 
 module RecordingStudioBilling
   class DummyFinancialAdapter < FakeFinancialAdapter
@@ -163,7 +165,22 @@ end
 
 RecordingStudioBilling.register_provider(:fake, RecordingStudioBilling::DummyFinancialAdapter.new)
 
+unless Rails.env.test?
+  stripe_test_credentials = DummyStripeTestCredentials.to_h
+  if stripe_test_credentials
+    RecordingStudioBilling.configure do |config|
+      config.stripe_credential_resolver = -> { DummyStripeTestCredentials.to_h }
+      config.stripe_trusted_origins = [ DummyStripeTestCredentials::RETURN_ORIGIN ]
+    end
+  end
+end
+
 Rails.application.config.to_prepare do
+  if DummyStripeTestCredentials.user_flow_enabled?
+    controller = "RecordingStudioBilling::CheckoutSelectionsController".constantize
+    controller.prepend(DummyStripeCheckoutExecution) unless controller < DummyStripeCheckoutExecution
+  end
+
   RecordingStudioBilling.configuration.feature_definitions = {
     "demo_priority_support" => {
       source: "catalogue", merge_rule: "replace", default: false, type: "boolean", meter_key: nil,